# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuthKit API
  version: 0.1.0-rc.1
  description: Static integration contract. Runtime documentation endpoints are intentionally disabled.
servers:
  - url: https://auth.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/users/me:
    get: &authGet
      responses: {'200': {description: Successful response}, '401': {description: Unauthorized}, '403': {description: Forbidden}}
    patch:
      requestBody: {$ref: '#/components/requestBodies/ProfileUpdate'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
    delete:
      requestBody: {$ref: '#/components/requestBodies/StepUp'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/consent:
    get: *authGet
  /api/v1/users/me/export:
    post:
      requestBody: {$ref: '#/components/requestBodies/StepUp'}
      responses: {'200': {$ref: '#/components/responses/UserDataExport'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/password:
    post:
      requestBody: {$ref: '#/components/requestBodies/PasswordChange'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/email-change:
    post:
      requestBody: {$ref: '#/components/requestBodies/EmailChange'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
    delete:
      requestBody: {$ref: '#/components/requestBodies/StepUp'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/sessions:
    get:
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: cursor, in: query, schema: {type: string}, description: Opaque single-use cursor with a five-minute lifetime}
      responses:
        '200':
          description: Bounded active-session page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionPageApiResponse'}
        '400': {$ref: '#/components/responses/Error'}
  /api/v1/users/me/sessions/{sessionId}:
    delete:
      parameters: [{$ref: '#/components/parameters/SessionId'}]
      requestBody: {$ref: '#/components/requestBodies/MfaOptionalVerification'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
  /api/v1/users/me/mfa:
    get: *authGet
  /api/v1/users/me/mfa/totp/enroll:
    post:
      requestBody: {$ref: '#/components/requestBodies/StepUp'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/mfa/totp/confirm:
    post:
      requestBody: {$ref: '#/components/requestBodies/MfaTotpConfirm'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/mfa/totp:
    delete:
      requestBody: {$ref: '#/components/requestBodies/MfaVerification'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/mfa/backup-codes:
    post:
      requestBody: {$ref: '#/components/requestBodies/MfaVerification'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/passkeys:
    get: *authGet
    post:
      requestBody: {$ref: '#/components/requestBodies/PasskeyRegistrationFinish'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/passkeys/options:
    post:
      requestBody: {$ref: '#/components/requestBodies/StepUpOptional'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/users/me/passkeys/{credentialId}:
    delete:
      parameters: [{$ref: '#/components/parameters/CredentialId'}]
      requestBody: {$ref: '#/components/requestBodies/StepUpOptional'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: First-party AuthKit access token with `token_use=first_party_access`, configured API audience, non-revoked key, and active server-side session. OAuth access tokens and ID tokens are rejected.
    oauthAccessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth access token with `token_use=oauth_access`, OAuth client audience, `client_id`, and server-side scope checks. ID tokens and first-party access tokens are rejected.
    refreshCookie:
      type: apiKey
      in: cookie
      name: Refresh-Token
      description: HttpOnly refresh cookie scoped to AuthKit refresh/logout routes.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: CSRF header that must echo the readable CSRF cookie on refresh and logout.
    oauthClientBasic:
      type: http
      scheme: basic
      description: OAuth client authentication for token, introspection, and revocation endpoints. Public clients may use approved body credentials where supported.
    workerToken:
      type: apiKey
      in: header
      name: X-Worker-Token
      description: Rotatable internal credential accepted only from explicitly trusted network peers.
  parameters:
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
...
    CredentialId: {name: credentialId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
...
    ProfileUpdate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProfileUpdateRequest'}}}}
    StepUp: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StepUpRequest'}}}}
    StepUpOptional: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/StepUpRequest'}}}}
    MfaOptionalVerification: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/MfaOptionalVerificationRequest'}}}}
...
    PasswordChange: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PasswordChangeRequest'}}}}
    EmailChange: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EmailChangeRequest'}}}}
    MfaTotpConfirm: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MfaTotpConfirmRequest'}}}}
    MfaVerification: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MfaVerificationRequest'}}}}
    PasskeyRegistrationFinish: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PasskeyRegistrationFinishRequest'}}}}
...
  headers:
    XRequestId:
      required: true
      schema: {type: string, format: uuid}
      description: Request correlation identifier; a valid inbound X-Request-ID is preserved, otherwise AuthKit generates one.
  responses:
    Success:
      description: Successful response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
    Error:
      description: Opaque error response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
...
    UserDataExport:
      description: Versioned data-subject export; credential and provider secrets are omitted
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/UserDataExportApiResponse'}}}
  schemas:
...
    ProfileUpdateRequest:
      type: object
      additionalProperties: false
      properties: {name: {type: string, maxLength: 100, nullable: true}}
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [newPassword]
      properties:
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        newPassword: {type: string, minLength: 12, maxLength: 128, writeOnly: true}
        mfaCode: {type: string, minLength: 6, maxLength: 32, nullable: true, writeOnly: true}
    EmailChangeRequest:
      type: object
      additionalProperties: false
      required: [newEmail, currentPassword]
      properties:
        newEmail: {type: string, format: email, maxLength: 255}
        currentPassword: {type: string, minLength: 1, maxLength: 128, writeOnly: true}
        mfaCode: {type: string, minLength: 6, maxLength: 32, nullable: true, writeOnly: true}
    MfaTotpConfirmRequest:
      type: object
      additionalProperties: false
      required: [credentialId, currentPassword, code]
      properties:
        credentialId: {type: string, format: uuid}
        currentPassword: {type: string, minLength: 1, maxLength: 128, writeOnly: true}
        code: {type: string, minLength: 6, maxLength: 8, writeOnly: true}
    MfaVerificationRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, code]
      properties:
        currentPassword: {type: string, minLength: 1, maxLength: 128, writeOnly: true}
        code: {type: string, minLength: 6, maxLength: 32, writeOnly: true}
    PasskeyRegistrationFinishRequest:
      type: object
      additionalProperties: false
      required: [challengeId, label, credentialJson]
      properties:
        challengeId: {type: string, format: uuid}
        label: {type: string, minLength: 1, maxLength: 100}
        credentialJson: {type: string, minLength: 1, maxLength: 20000, writeOnly: true}
...
    ApiResponse:
      type: object
      additionalProperties: false
      required: [timestamp, requestId]
      properties:
        data: {}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        code: {type: string, pattern: '^[a-z0-9_]+$', nullable: true, description: Stable machine-readable error code}
        requestId: {type: string, minLength: 1, maxLength: 128}
    UserDataExportApiResponse:
      type: object
      additionalProperties: false
      required: [data, timestamp, requestId]
      properties:
        data: {$ref: '#/components/schemas/UserDataExportResponse'}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        code: {type: string, pattern: '^[a-z0-9_]+$', nullable: true}
        requestId: {type: string, minLength: 1, maxLength: 128}
    UserDataExportResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, generatedAt, profile, consent, consentHistory, oauthConsents, credentials, sessions, deletion, securityEvents]
      properties:
        schemaVersion: {type: string, enum: [authkit-user-data-export/v1]}
        generatedAt: {type: string, format: date-time}
        profile:
          type: object
          additionalProperties: false
          required: [id, email, role, tenantId, emailConfirmed, accountState]
          properties:
            id: {type: string, format: uuid}
            email: {type: string, format: email}
            name: {type: string, nullable: true}
            role: {type: string, enum: [USER, PLATFORM_ADMIN]}
            tenantId: {type: string, format: uuid, description: Opaque per-person partition}
            emailConfirmed: {type: boolean}
            accountState: {type: string, enum: [ACTIVE, SUSPENDED, DELETION_PENDING, ANONYMIZED]}
            suspendedAt: {type: string, format: date-time, nullable: true}
            suspensionReason: {type: string, nullable: true}
            pendingEmail: {type: string, format: email, nullable: true}
            emailChangeRequestedAt: {type: string, format: date-time, nullable: true}
            emailChangeExpiresAt: {type: string, format: date-time, nullable: true}
        consent:
          type: object
          additionalProperties: false
          required: [termsAccepted, privacyPolicyAccepted]
          properties:
            termsAccepted: {type: boolean}
            privacyPolicyAccepted: {type: boolean}
            termsVersion: {type: string, nullable: true}
            privacyPolicyVersion: {type: string, nullable: true}
            consentAcceptedAt: {type: string, format: date-time, nullable: true}
            lawfulBasis: {type: string, nullable: true}
        consentHistory:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [acceptedAt, recordedAt, eventHash]
            properties:
              termsVersion: {type: string, nullable: true}
              privacyPolicyVersion: {type: string, nullable: true}
              lawfulBasis: {type: string, nullable: true}
              acceptedAt: {type: string, format: date-time}
              recordedAt: {type: string, format: date-time}
              eventHash: {type: string, description: Integrity evidence; not a secret}
        oauthConsents:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [clientId, scopes, grantedAt]
            properties:
              clientId: {type: string}
              scopes: {type: array, items: {type: string}}
              grantedAt: {type: string, format: date-time}
              revokedAt: {type: string, format: date-time, nullable: true}
        credentials:
          type: object
          additionalProperties: false
          required: [localPasswordConfigured, passwordHistory, totp, passkeys, socialIdentities]
          properties:
            localPasswordConfigured: {type: boolean, description: Boolean only; password hash is never exported}
            passwordHistory: {type: array, items: {type: object, additionalProperties: false, required: [recordedAt], properties: {recordedAt: {type: string, format: date-time}}}}
            totp: {type: array, items: {type: object, additionalProperties: false, required: [id, createdAt], properties: {id: {type: string, format: uuid}, createdAt: {type: string, format: date-time}, confirmedAt: {type: string, format: date-time, nullable: true}, disabledAt: {type: string, format: date-time, nullable: true}}}}
            passkeys: {type: array, items: {type: object, additionalProperties: false, required: [id, discoverable, signatureCount, createdAt], properties: {id: {type: string, format: uuid}, label: {type: string, nullable: true}, transports: {type: string, nullable: true}, discoverable: {type: boolean}, signatureCount: {type: integer, format: int64, minimum: 0}, createdAt: {type: string, format: date-time}, lastUsedAt: {type: string, format: date-time, nullable: true}, disabledAt: {type: string, format: date-time, nullable: true}}}}
            socialIdentities: {type: array, items: {type: object, additionalProperties: false, required: [id, providerKey, issuer, emailVerified, createdAt], properties: {id: {type: string, format: uuid}, providerKey: {type: string}, issuer: {type: string, format: uri}, emailAtLink: {type: string, format: email, nullable: true}, emailVerified: {type: boolean}, createdAt: {type: string, format: date-time}, lastLoginAt: {type: string, format: date-time, nullable: true}}}}
        sessions:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [sessionId, createdAt, lastSeenAt, expiresAt, initialAmr, userAgentSummary, creationIpMasked, lastIpMasked]
            properties:
              sessionId: {type: string, format: uuid}
              createdAt: {type: string, format: date-time}
              lastSeenAt: {type: string, format: date-time}
              expiresAt: {type: string, format: date-time}
              initialAmr: {type: array, items: {type: string}}
              userAgentSummary: {type: string}
              deviceLabel: {type: string, nullable: true}
              creationIpMasked: {type: string}
              lastIpMasked: {type: string}
        deletion:
          type: object
          additionalProperties: false
          properties:
            deletionRequestedAt: {type: string, format: date-time, nullable: true}
            deletedAt: {type: string, format: date-time, nullable: true}
            anonymizedAt: {type: string, format: date-time, nullable: true}
        securityEvents:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [occurredAt, type, outcome, severity]
            properties:
              occurredAt: {type: string, format: date-time}
              type: {type: string}
              outcome: {type: string, enum: [SUCCESS, FAILURE, DENIED]}
              severity: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL]}
              emailMasked: {type: string, nullable: true}
              clientIpMasked: {type: string, nullable: true}
              requestMethod: {type: string, nullable: true}
              requestPath: {type: string, nullable: true}
              reason: {type: string, nullable: true}
    SessionPageApiResponse:
      type: object
      required: [data, timestamp, requestId]
      properties:
        data: {$ref: '#/components/schemas/SessionPageResponse'}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        requestId: {type: string}
    SessionPageResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/SessionResponse'}
        nextCursor: {type: string, nullable: true, description: Opaque single-use cursor}
    SessionResponse:
      type: object
      required: [sessionId, createdAt, lastSeenAt, expiresAt, current, initialAmr, userAgentSummary, creationIpMasked, lastIpMasked]
      properties:
        sessionId: {type: string, format: uuid, description: Opaque public session identifier; never a JWT JTI or refresh secret}
        createdAt: {type: string, format: date-time}
        lastSeenAt: {type: string, format: date-time, description: 'Approximate activity, persisted at a bounded frequency'}
        expiresAt: {type: string, format: date-time}
        current: {type: boolean}
        initialAmr: {type: array, items: {type: string}}
        userAgentSummary: {type: string, maxLength: 200}
        deviceLabel: {type: string, maxLength: 80, nullable: true}
        creationIpMasked: {type: string}
        lastIpMasked: {type: string}
...
    StepUpRequest:
      type: object
      additionalProperties: false
      properties:
        currentPassword: {type: string, maxLength: 128, nullable: true}
        mfaCode: {type: string, minLength: 6, maxLength: 32, nullable: true}
    MfaOptionalVerificationRequest:
      type: object
      additionalProperties: false
      properties:
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        code: {type: string, minLength: 6, maxLength: 32, nullable: true, writeOnly: true}
...